feat(ready-page): ladder/chanllenge game type ready page
